pyproject.toml
[project]
name = "qi"
version = "0.1.0"
description = "Qi - Better Workflow for Contract-Based Development"
requires-python = ">=3.10"
authors = [
    {name = "Nick Song", email = "nsong81@gmail.com"}
]
dependencies = [
    "typer>=0.15.1",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "openapi-spec-validator>=0.7.1",
    "apispec>=6.3.0",
]

[project.urls]
Homepage = "https://github.com/nicksenap/qi"
Repository = "https://github.com/nicksenap/qi"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["qi"]

[project.scripts]
qi = "qi.cli:app"

[tool.pytest.ini_options]
addopts = "-v --cov=qi --cov-report=term-missing"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "PL",  # pylint
    "RUF", # ruff-specific rules
    "TCH", # type-checking
    "TID", # tidy imports
    "ERA", # eradicate
]
ignore = [
    "B008",  # Do not perform function calls in argument defaults (needed for Typer)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004"]  # Ignore magic numbers in tests
